+++ src/components/fields/index.ts
@@ -1,12 +1,12 @@
-export { ReferenceField } from "./referenceField";
-export { TextField } from "./textField";
-export { TagField } from "./tagField";
-export { EmailField } from "./emailField";
-export { ImageField } from "./imageField";
-export { BooleanField } from "./booleanField";
-export { DateField } from "./dateField";
-export { FileField } from "./fileField";
-export { UrlField } from "./urlField";
-export { NumberField } from "./numberField";
-export { MarkdownField } from "./markdownField";
-export { RichtextField } from "./richtextField";
+export { ReferenceField } from "./reference";
+export { TextField } from "./text";
+export { TagField } from "./tag";
+export { EmailField } from "./email";
+export { ImageField } from "./image";
+export { BooleanField } from "./boolean";
+export { DateField } from "./date";
+export { FileField } from "./file";
+export { UrlField } from "./url";
+export { NumberField } from "./number";
+export { MarkdownField } from "./markdown";
+export { RichtextField } from "./richtext";
